Laborator 2

+++ Laborator2/problema2/Student.cpp
@@ -15,17 +15,17 @@ void Student::setName(const char* name)
 	strcpy(this->name, name);
 }
 
-void Student::setMathGrade(short math_grade)
+void Student::setMathGrade(float math_grade)
 {
 	this->math_grade = math_grade;
 }
 
-void Student::setEnglishGrade(short english_grade)
+void Student::setEnglishGrade(float english_grade)
 {
 	this->english_grade = english_grade;
 }
 
-void Student::setHistoryGrade(short history_grade)
+void Student::setHistoryGrade(float history_grade)
 {
 	this->history_grade = history_grade;
 }
@@ -35,17 +35,17 @@ char* Student::getName()
 	return this->name;
 }
 
-short Student::getMathGrade()
+float Student::getMathGrade()
 {
 	return this->math_grade;
 }
 
-short Student::getEnglishGrade()
+float Student::getEnglishGrade()
 {
 	return this->english_grade;
 }
 
-short Student::getHistoryGrade()
+float Student::getHistoryGrade()
 {
 	return this->history_grade;
 }
