+++ Laborator2/problema2/Student.h
@@ -4,19 +4,19 @@ class Student
 {
 private:
 	char name[100];
-	short math_grade;
-	short english_grade;
-	short history_grade;
+	float math_grade;
+	float english_grade;
+	float history_grade;
 public:
 	Student();
 	void setName(const char* name);
-	void setMathGrade(short math_grade);
-	void setEnglishGrade(short english_grade);
-	void setHistoryGrade(short history_grade);
+	void setMathGrade(float math_grade);
+	void setEnglishGrade(float english_grade);
+	void setHistoryGrade(float history_grade);
 	char* getName();
-	short getMathGrade();
-	short getEnglishGrade();
-	short getHistoryGrade();
+	float getMathGrade();
+	float getEnglishGrade();
+	float getHistoryGrade();
 	float getAverage();
 };
